refactor: 도메인 enum 타입을 domain/enums 패키지로 분리
- DailyStockStatus, OrderStatus, PaymentStatus를
  com.chaltteok.core.domain → com.chaltteok.core.domain.enums로 이동
- 참조하는 모든 모듈(deal-core, deal-consumer, deal-api-user) import 업데이트
