+++ deal-core/src/main/kotlin/com/chaltteok/core/domain/Order.kt
@@ -1,5 +1,6 @@
 package com.chaltteok.core.domain
 
+import com.chaltteok.core.domain.enums.OrderStatus
 import jakarta.persistence.*
 import java.time.LocalDateTime
 import java.util.*
