+++ deal-core/src/main/kotlin/com/chaltteok/core/domain/enums/DailyStockStatus.kt
@@ -1,4 +1,4 @@
-package com.chaltteok.core.domain
+package com.chaltteok.core.domain.enums
 
 enum class DailyStockStatus {
     OPEN, SOLD_OUT, CLOSED
