+++ deal-core/src/main/kotlin/com/chaltteok/core/domain/enums/OrderStatus.kt
@@ -1,4 +1,4 @@
-package com.chaltteok.core.domain
+package com.chaltteok.core.domain.enums
 
 enum class OrderStatus {
     PENDING, COMPLETED, CANCELLED, FAILED
