+++ deal-core/src/main/kotlin/com/chaltteok/core/domain/enums/PaymentStatus.kt
@@ -1,4 +1,4 @@
-package com.chaltteok.core.domain
+package com.chaltteok.core.domain.enums
 
 enum class PaymentStatus {
     READY, SUCCESS, FAILED, CANCELLED
